backend/debt-away/api/requirements.txt
boto3
requests
fastapi
mangum
plaid-python==27.0.0